+++ src/lib/pages/evm-contract-details/components/EvmContractDetailsTop.tsx
@@ -1,18 +1,23 @@
-import { Flex, Heading, Stack, Text } from "@chakra-ui/react";
+import { Button, Flex, Grid, Heading, Stack, Text } from "@chakra-ui/react";
 
-import { useConvertHexAddress } from "lib/app-provider";
+import { useConvertHexAddress, useInternalNavigate } from "lib/app-provider";
 import { CopyLink } from "lib/components/CopyLink";
 import { CustomIcon } from "lib/components/icon";
 import { TotalValue } from "lib/components/TotalValue";
 import type { HexAddr20 } from "lib/types";
+import { TabIndex } from "../types";
+import { getInteractTabsIndex } from "../utils";
 
 interface EvmContractDetailsTopProps {
   contractAddress: HexAddr20;
+  isVerified: boolean;
 }
 
 export const EvmContractDetailsTop = ({
   contractAddress,
+  isVerified,
 }: EvmContractDetailsTopProps) => {
+  const navigate = useInternalNavigate();
   const { convertHexWalletAddress } = useConvertHexAddress();
 
   return (
@@ -36,18 +41,64 @@ export const EvmContractDetailsTop = ({
           <Text variant="body2" fontWeight={500} color="text.dark">
             Contract Address:
           </Text>
-          <CopyLink
-            value={contractAddress}
-            amptrackSection="contract_top"
-            type="contract_address"
-          />
+          <Flex alignItems="center">
+            <CopyLink
+              value={contractAddress}
+              amptrackSection="contract_top"
+              type="contract_address"
+            />
+            {isVerified && (
+              <CustomIcon
+                name="verification-solid"
+                boxSize={4}
+                color="secondary.main"
+              />
+            )}
+          </Flex>
         </Flex>
       </Stack>
-      <TotalValue
-        address={convertHexWalletAddress(contractAddress)}
-        label="Total Value"
-        isCompact
-      />
+      <Stack gap={4}>
+        {isVerified && (
+          <Grid gap={2} templateColumns={{ md: "repeat(2, 1fr)", base: "1fr" }}>
+            <Button
+              variant="outline-white"
+              onClick={() =>
+                navigate({
+                  pathname: "/evm-contracts/[contractAddress]/[tab]",
+                  query: {
+                    contractAddress,
+                    tab: TabIndex.ReadWrite,
+                    selectedType: getInteractTabsIndex(true, false),
+                  },
+                })
+              }
+            >
+              Read
+            </Button>
+            <Button
+              onClick={() =>
+                navigate({
+                  pathname: "/evm-contracts/[contractAddress]/[tab]",
+                  query: {
+                    contractAddress,
+                    tab: TabIndex.ReadWrite,
+                    selectedType: getInteractTabsIndex(false, false),
+                  },
+                })
+              }
+              variant="outline-white"
+              display={{ base: "none", md: "block" }}
+            >
+              Write
+            </Button>
+          </Grid>
+        )}
+        <TotalValue
+          address={convertHexWalletAddress(contractAddress)}
+          label="Total Value"
+          isCompact
+        />
+      </Stack>
     </Flex>
   );
 };